backend/requirements.txt
accelerate==1.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asteroid-filterbanks==0.4.0
atomicwrites==1.4.0
audioread==3.0.1
av==14.4.0
coloredlogs==15.0.1
colorlog==6.9.0
ctranslate2==4.6.0
docopt==0.6.2
einops==0.8.1
enum34==1.1.10
fastapi==0.115.12
faster-whisper==1.1.1
flask-cors==6.0.0
flatbuffers==25.2.10
halo==0.0.31
hf-xet==1.1.2
huggingface-hub==0.32.4
humanfriendly==10.0
HyperPyYAML==1.2.2
jsonpointer==2.1
julius==0.2.7
librosa==0.11.0
lightning==2.5.1.post0
lightning-utilities==0.14.3
log-symbols==0.0.14
matplotlib==3.9.2
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.22.0
openai-whisper==20240930
openwakeword==0.6.0
optuna==4.4.0
pooch==1.8.2
primePy==1.3
protobuf==4.25.3
pvporcupine==1.9.5
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1-modules==0.2.8
PyAudio==0.2.14
pydub==0.25.1
pyls-spyder==0.4.0
PyQt5==5.15.10
PyQtWebEngine==5.15.6
python-multipart==0.0.20
pytorch-lightning==2.5.1.post0
pytorch-metric-learning==2.8.1
realtimestt==0.3.104
requests-oauthlib==2.0.0
Resemblyzer==0.1.4
safetensors==0.5.3
schedule==1.2.2
scipy==1.15.2
sentencepiece==0.2.0
setuptools==75.1.0
soundfile==0.13.1
soxr==0.5.0.post1
speechbrain==1.0.3
spinners==0.0.24
starlette==0.46.2
sympy==1.14.0
tensorboardX==2.6.4
termcolor==3.1.0
tiktoken==0.9.0
tk==0.1.0
tokenizers==0.21.1
torch==2.7.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.7.0
torchmetrics==1.7.3
transformers==4.52.4
tweepy==4.15.0
typer==0.16.0
typing==3.7.4.3
uvicorn==0.34.3
webrtcvad==2.0.10
webrtcvad-wheels==2.0.14
websockets==15.0.1
wheel==0.44.0
whisper==1.1.10
sounddevice==0.5.2
numpy
SpeechRecognition==3.14.3
pocketsphinx==5.0.4